# [TraceQL] Performance increase for simple queries