--- a/tempodb/encoding/vparquet4/block_traceql.go
+++ b/tempodb/encoding/vparquet4/block_traceql.go
@@ -2977,27 +2977,19 @@ func (c *instrumentationCollector) KeepGroup(res *parquetquery.IteratorResult) b
 		}
 	}
 
-	// Second pass. Update and further filter the spans
-	spans = res.OtherEntries[:0]
-	for _, e := range res.OtherEntries {
-		span, ok := e.Value.(*span)
-		if !ok {
-			continue
+	// Second pass. Update spans with instrumentation attributes.
+	if len(c.instrumentationAttrs) > 0 {
+		for _, e := range res.OtherEntries {
+			span, ok := e.Value.(*span)
+			if !ok {
+				continue
+			}
+
+			// Copy scope-level attributes to the span
+			// If the span already has an entry for this attribute it
+			// takes precedence (can be nil to indicate no match)
+			span.setInstrumentationAttrs(c.instrumentationAttrs)
 		}
-
-		// Copy scope-level attributes to the span
-		// If the span already has an entry for this attribute it
-		// takes precedence (can be nil to indicate no match)
-		span.setInstrumentationAttrs(c.instrumentationAttrs)
-		spans = append(spans, e)
-
-	}
-
-	// pass up to resource collector
-	res.OtherEntries = spans
-	// Throw out batches without any remaining spans
-	if len(res.OtherEntries) == 0 {
-		return false
 	}
 
 	res.Entries = res.Entries[:0]
@@ -3065,30 +3057,31 @@ func (c *batchCollector) KeepGroup(res *parquetquery.IteratorResult) bool {
 	}
 
 	// Second pass. Update and further filter the spans
-	spans = res.OtherEntries[:0]
-	for _, e := range res.OtherEntries {
-		span, ok := e.Value.(*span)
-		if !ok {
-			continue
-		}
-
-		// Copy resource-level attributes to the span
-		// If the span already has an entry for this attribute it
-		// takes precedence (can be nil to indicate no match)
-		span.setResourceAttrs(c.resAttrs)
-
-		if c.requireAtLeastOneMatchOverall {
-			// Skip over span if it didn't meet minimum criteria
-			if span.attributesMatched() == 0 {
-				putSpan(span)
+	if len(c.resAttrs) > 0 || c.requireAtLeastOneMatchOverall {
+		spans = res.OtherEntries[:0]
+		for _, e := range res.OtherEntries {
+			span, ok := e.Value.(*span)
+			if !ok {
 				continue
 			}
+
+			// Copy resource-level attributes to the span
+			// If the span already has an entry for this attribute it
+			// takes precedence (can be nil to indicate no match)
+			span.setResourceAttrs(c.resAttrs)
+
+			if c.requireAtLeastOneMatchOverall {
+				// Skip over span if it didn't meet minimum criteria
+				if span.attributesMatched() == 0 {
+					putSpan(span)
+					continue
+				}
+			}
+
+			spans = append(spans, e)
 		}
-
-		spans = append(spans, e)
-
+		res.OtherEntries = spans
 	}
-	res.OtherEntries = spans
 
 	// Throw out batches without any remaining spans
 	if len(res.OtherEntries) == 0 {
@@ -3163,9 +3156,11 @@ func (c *traceCollector) KeepGroup(res *parquetquery.IteratorResult) bool {
 	}
 
 	// loop over all spans and add the trace-level attributes
-	for _, s := range finalSpanset.Spans {
-		s := s.(*span)
-		s.setTraceAttrs(c.traceAttrs)
+	if len(c.traceAttrs) > 0 {
+		for _, s := range finalSpanset.Spans {
+			s := s.(*span)
+			s.setTraceAttrs(c.traceAttrs)
+		}
 	}
 
 	if numServiceStats > 0 {